+++ example/data_source_virtual_environment_access_groups.tf
@@ -3,6 +3,6 @@ data "proxmox_virtual_environment_access_groups" "example" {}
 output "data_proxmox_virtual_environment_access_groups_example" {
   value = "${map(
     "comments", data.proxmox_virtual_environment_access_groups.example.comments,
-    "ids", data.proxmox_virtual_environment_access_groups.example.ids,
+    "group_ids", data.proxmox_virtual_environment_access_groups.example.group_ids,
   )}"
 }