+++ CHANGELOG.md
@@ -4,5 +4,8 @@ FEATURES:
 
 * **New Data Source:** `proxmox_virtual_environment_access_group`
 * **New Data Source:** `proxmox_virtual_environment_access_groups`
+* **New Data Source:** `proxmox_virtual_environment_access_role`
+* **New Data Source:** `proxmox_virtual_environment_access_roles`
 * **New Data Source:** `proxmox_virtual_environment_version`
 * **New Resource:** `proxmox_virtual_environment_access_group`
+* **New Resource:** `proxmox_virtual_environment_access_role`